requirements.txt
Flask==3.1.2
Flask-Cors==4.0.1
gunicorn==22.0.0

discord.py==2.4.0

python-dotenv==1.0.1
